src/api/requirements.txt
fastapi[standard]==0.115.6
aiohttp==3.11.9
pydantic==2.9.2
pydantic-settings==2.6.1
nest_asyncio==1.6.0
fastapi-profiler==1.4.1
semantic-kernel[azure]==1.21.2
requests==2.32.3
opentelemetry-exporter-otlp==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
pyyaml==6.0.2
azure-monitor-opentelemetry-exporter==1.0.0b33
azure-identity==1.20.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-monitor-opentelemetry==1.6.5